% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_sql.R
\name{csv_to_sql}
\alias{csv_to_sql}
\title{Converts directory or list of .csv files into an SQLite database}
\usage{
csv_to_sql(x, ...)
}
\arguments{
\item{x}{list of files or directory containing files}

\item{...}{additional arguments to be passed to \code{csv_to_sqlite_x}}
}
\description{
Produces an SQLite database from either a directory or a list of .csv files
}

