% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_sqlite_list_.R
\name{csv_to_sqlite_list_}
\alias{csv_to_sqlite_list_}
\title{Converts directory of .csv files to an SQLite database}
\usage{
csv_to_sqlite_list_(files, dir = NULL, db_name = "db",
  single_table = FALSE, table_name = "exp")
}
\arguments{
\item{files}{list of files}

\item{dir}{Directory to save database in, otherwise will save in the
current working directory.}

\item{db_name}{Name given to the database.}

\item{single_table}{Whether to place all .csv files in a single table,
otherwise will produce a table per file.}

\item{table_name}{Only relevant if `single_table = TRUE`: what to call the table.}
}
\description{
Called by `csv_to_sql`, can also be used directly. If given a
list of .csv files, will generate an SQLite database containing either all the 
files in a single table, or a table for each file.
}

