% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_sqlite_path_.R
\name{csv_to_sqlite_path_}
\alias{csv_to_sqlite_path_}
\title{Converts directory of .csv files to an SQLite database}
\usage{
csv_to_sqlite_path_(path, dir = NULL, db_name = "db",
  single_table = FALSE, table_name = "exp")
}
\arguments{
\item{path}{Filepath to directory containing .csv files}

\item{dir}{Directory to save database in, otherwise will save in the
filepath given in the `path` argument.}

\item{db_name}{Name given to the database.}

\item{single_table}{Whether to place all .csv files in a single table,
otherwise will produce a table per file.}

\item{table_name}{Only useful if `single_table = TRUE`, what to call the table.}
}
\description{
Called by `csv_to_sql`, can also be used directly. IFf given a
filepath, will generate an sqlite database containing either all the 
files in a single table, or a table for each file.
}

