% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_merge_objects.R
\name{featuredata_cols}
\alias{featuredata_cols}
\title{Get featuredata columns}
\usage{
featuredata_cols(table, value = TRUE)
}
\arguments{
\item{table}{dataframe of table within database}

\item{value}{If \code{TRUE} will return the column names, if \code{FALSE}
will return the indicies}
}
\description{
Returns the feature data columns from a dataframe, feature data is classed
 as anything that is not metadata
}

