% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_merge_objects.R
\name{metadata_cols}
\alias{metadata_cols}
\title{Get metadata columns}
\usage{
metadata_cols(table, value = TRUE)
}
\arguments{
\item{table}{dataframe or table within database}

\item{value}{If \code{TRUE} will return the column names, if \code{FALSE}
will return the indicies}
}
\description{
Returns metadata columns from dataframe, by grep'ing columns beginning
with 'Metadata_'
}

