% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sql_merge_objects.R
\name{sql_merge_objects}
\alias{sql_merge_objects}
\title{Merge object data}
\usage{
sql_merge_objects(db, sep = "_")
}
\arguments{
\item{db}{database connection}

\item{sep}{seperator between pipeline and object}
}
\description{
Function to merge together multiple objects from a pipeline into a single
table and prefix the column names with the object label.
}

