% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_merge_objects.R
\name{sql_merge_objects}
\alias{sql_merge_objects}
\title{Merge object data}
\usage{
sql_merge_objects(db, metrics = NULL, sep = "_")
}
\arguments{
\item{db}{database connection}

\item{metrics}{helper filepath to file detailing pipelines and objects}

\item{sep}{separator to separate pipeline and object in the filename
only used if a metric file is not supplied}
}
\description{
Function to merge together multiple objects from a pipeline into a single
table and prefix the column names with the object label.
}

